Add feed item for 'account created'
