+++ src/components/WalletEventItem.vue
@@ -1,33 +1,5 @@
 <template>
-  <b-container class="p-0">
-    <b-card no-body class="px-2 pt-2">
-      <!-- User -->
-      <b-row class="m-0" >
-        <!-- Picture -->
-        <b-col cols="auto" class="p-0">
-          <b-avatar size="3rem" :src="item.owner.picture" :to="`/users/${item.owner.username}`"/>
-        </b-col>
-
-        <b-col style="line-height: 1.1rem" class="pl-4">
-          <!-- Username -->
-          <b-row>
-            <b-link :to="`/users/${item.owner.username}`" class="text-dark text-decoration-none">
-              <strong class="text-truncate"> {{ item.owner.username }} </strong>
-            </b-link>
-          </b-row>
-          <!-- Title -->
-          <b-row>
-          <span v-if="item.owner.title" class="text-secondary" style="font-size: 0.85rem">
-            {{ item.owner.title }}
-          </span>
-          </b-row>
-          <!-- Timestamp -->
-          <b-row class="text-secondary text-truncate" style="font-size: 0.75rem">
-            {{ timestamp }}
-          </b-row>
-        </b-col>
-      </b-row>
-
+  <div>
       <!-- Event -->
       <b-row class="pt-1">
         <b-col>
@@ -197,31 +169,14 @@
         </b-col>
       </b-row>
 
-      <!-- Reaction button -->
-      <b-card-footer class="pt-0 m-0 px-3 pb-0" footer-bg-variant="white">
-        <b-row align-v="center">
-          <!-- Like -->
-          <b-button v-if="!this.like" @click="likePost" variant="white" size="md">
-            <font-awesome-icon :icon="['far', 'thumbs-up']" />
-            <span> {{ likeCount }} </span>
-          </b-button>
-
-          <!-- Unlike -->
-          <b-button v-else @click="unlikePost" variant="white" size="md">
-            <font-awesome-icon :icon="['fas', 'thumbs-up']" class="text-primary" />
-            <span> {{ likeCount }} </span>
-          </b-button>
-        </b-row>
-      </b-card-footer>
-    </b-card>
-  </b-container>
+    </div>
 </template>
 
 <script>
 import client from '@/commons/client.api'
 
 export default {
-  name: 'FeedItem',
+  name: 'WalletEventItem',
 
   props: {
     item: {
@@ -230,22 +185,16 @@ export default {
     }
   },
 
-  data: () => ({
-    like: false,
-    likeCount: 0
-  }),
-
   computed: {
     text () {
       switch (this.item.type) {
         case 'NEW_WALLET':
           return 'criou uma'
-        default:
+        case 'WALLET_ASSETS_EDIT':
           return 'atualizou sua'
+        default:
+          return ''
       }
-    },
-    timestamp () {
-      return new Date(this.item.timestamp).toLocaleString('pt-BR').substring(0, 16)
     },
     increases () {
       return this.item.data.changes.filter(evt => evt.change === 'INC')
@@ -264,22 +213,8 @@ export default {
   methods: {
     thumb (ticker) {
       return client.utils.thumbUrl(ticker)
-    },
-    async likePost () {
-      await client.post.likePost(this.item.id)
-      this.like = true
-      this.likeCount += 1
-    },
-    async unlikePost () {
-      await client.post.unlikePost(this.item.id)
-      this.like = false
-      this.likeCount -= 1
     }
-  },
-
-  created () {
-    this.like = this.item.like
-    this.likeCount = this.item.likeCount
   }
+
 }
 </script>
