+++ src/views/FeedView.vue
@@ -20,7 +20,7 @@
           <!-- 5 items of feed -->
           <b-row v-for="item of feed.slice(0,5)" :key="item.id" class="mt-3">
             <b-col>
-              <FeedItem :item="item" />
+              <EventItem :item="item" />
             </b-col>
           </b-row>
           <!-- suggested friends -->
@@ -32,7 +32,7 @@
           <!-- rest of feed (that grows as scrolls down) -->
           <b-row v-for="item of feed.slice(5)" :key="item.id" class="mt-3">
             <b-col>
-              <FeedItem :item="item" />
+              <EventItem :item="item" />
             </b-col>
           </b-row>
 
@@ -56,7 +56,7 @@
 import client from '@/commons/client.api'
 
 import Navbar from '@/components/Navbar'
-import FeedItem from '@/components/FeedItem'
+import EventItem from '@/components/EventItem'
 import InviteFriends from '@/components/InviteFriends'
 import NewUser from '@/components/NewUser'
 import SuggestedFriendsMobile from '@/components/SuggestedFriendsMobile'
@@ -67,7 +67,7 @@ export default {
 
   components: {
     Navbar,
-    FeedItem,
+    EventItem,
     InviteFriends,
     NewUser,
     Loading,
